+++ changelog.md
@@ -22,6 +22,7 @@ While all _Major_ releases should be reviewed, our only _large_ releases are **v
 - **[new]** - .nouns('spencer') shorthand as an if-match
 - **[change]** - "[do] you .." etc now #QuestionWord
 - **[new]** - add #Hyphenated tag
+- **[fix]** - improved Auxiliary verb tagging
 -->
 
 #### 14.7.1